+++ mobile/android/android-components/components/feature/ipprotection/src/main/java/mozilla/components/feature/ipprotection/store/Reducers.kt
@@ -160,6 +160,10 @@ internal fun iPProtectionReducer(
         state.copy(selectedCountryCode = action.code)
     }
 
+    is IPProtectionAction.CurrentSiteExcludedChanged -> {
+        state.copy(currentSiteExcluded = action.excluded)
+    }
+
     is InternalAction -> internalReducer(state, action)
 }