+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/settings/IPProtectionFragment.kt
@@ -94,6 +94,12 @@ class IPProtectionFragment : Fragment(), SystemInsetsPaddedFragment {
                             .actionIpProtectionFragmentToIpProtectionLocationsFragment(),
                     )
                 },
+                onManageWebsitesClick = {
+                    findNavController().navigate(
+                        IPProtectionFragmentDirections
+                            .actionIpProtectionFragmentToIpProtectionExceptionsFragment(),
+                    )
+                },
                 onVpnToggle = { enabled -> toggleVpn(enabled, state.selectedCountryCode) },
                 onLearnMoreClick = {
                     Vpn.settingsLearnMoreTapped.record(NoExtras())